requirements.txt
# Librería principal para AMQP/RabbitMQ
pika==1.3.2

# Para manejo de datos y serialización
python-dotenv==1.0.0
json5==0.9.14

# Para testing (opcional)
pytest==7.4.3
pytest-cov==4.1.0

# Para logging mejorado (opcional)
structlog==24.1.0

# Para manejo de fechas y tiempo (opcional)
python-dateutil==2.8.2

# Para validación de datos (opcional)
pydantic==2.5.2

# Para desarrollo y debugging (opcional)
black==23.11.0
flake8==6.1.0
isort==5.12.0